Arreglando validacion global Exception handler..

+++ RealStateApp.Core.Application/Features/Agents/Commands/ChangeStatus/ChangeAgentStatusCommand.cs
@@ -29,7 +29,7 @@ namespace RealStateApp.Core.Application.Features.Agents.Commands.ChangeStatus
 
             var operationStatus= await _userService.SetStatus(request.Id, request.Status);
 
-            if (operationStatus == false) throw new ApiException("Entity Not found", HttpStatusCode.NotFound);
+            if (operationStatus == false) throw new ApiException("Entity Not found", (int)HttpStatusCode.NotFound);
             return Unit.Value;
         }
     }
